[JENKINS-45804] - Refactor and remove OrganizationUtil. getOrganizationFromURL is now in OrganizationImpl and is restricted.
