+++ blueocean-rest-impl/src/main/java/io/jenkins/blueocean/service/embedded/UserStatePreloader.java
@@ -28,8 +28,8 @@ import hudson.model.User;
 import io.jenkins.blueocean.commons.PageStatePreloader;
 import io.jenkins.blueocean.commons.stapler.Export;
 import io.jenkins.blueocean.rest.model.BlueOrganization;
+import io.jenkins.blueocean.service.embedded.rest.OrganizationImpl;
 import io.jenkins.blueocean.service.embedded.rest.UserImpl;
-import io.jenkins.blueocean.service.embedded.util.OrganizationUtil;
 import net.sf.json.JSONObject;
 
 import java.io.IOException;
@@ -62,7 +62,7 @@ public class UserStatePreloader extends PageStatePreloader {
     @Override
     public String getStateJson() {
 
-        BlueOrganization organization = OrganizationUtil.getOrganization(OrganizationUtil.getOrganizationNameFromURL(), true);
+        BlueOrganization organization = OrganizationImpl.getOrganizationFromURL();
 
         try {
             User currentUser = User.current();
