+++ blueocean-rest-impl/src/main/java/io/jenkins/blueocean/service/embedded/rest/FreeStyleRunImpl.java
@@ -12,7 +12,6 @@ import io.jenkins.blueocean.rest.model.BlueChangeSetEntry;
 import io.jenkins.blueocean.rest.model.BlueRun;
 import io.jenkins.blueocean.rest.model.Container;
 import io.jenkins.blueocean.rest.model.Containers;
-import io.jenkins.blueocean.service.embedded.util.OrganizationUtil;
 
 import org.kohsuke.stapler.QueryParameter;
